fix wrong lookup of image values, fix rounding errors in pixel constructor

+++ src/neighbors.cpp
@@ -10,26 +10,26 @@ std::vector<Pixel> getNeighbors(const Pixel& p,
     neighbors.reserve(static_cast<size_t>(mode));
 
     if (p.row > 0)
-        neighbors.emplace_back(p.col, p.row - 1, img.at<double>(p.row - 1, p.col));
+        neighbors.emplace_back(p.col, p.row - 1, img.at<float>(p.row - 1, p.col));
     if (p.col > 0)
-        neighbors.emplace_back(p.col - 1, p.row, img.at<double>(p.row, p.col - 1));
+        neighbors.emplace_back(p.col - 1, p.row, img.at<float>(p.row, p.col - 1));
 
     if (static_cast<int>(mode) > 2) {
         if (p.row < img.rows - 1)
-            neighbors.emplace_back(p.col, p.row + 1, img.at<double>(p.row + 1, p.col));
+            neighbors.emplace_back(p.col, p.row + 1, img.at<float>(p.row + 1, p.col));
         if (p.col < img.cols - 1)
-            neighbors.emplace_back(p.col + 1, p.row, img.at<double>(p.row, p.col + 1));
+            neighbors.emplace_back(p.col + 1, p.row, img.at<float>(p.row, p.col + 1));
     }
 
     if (static_cast<int>(mode) > 4) {
         if (p.row > 0 && p.col < img.cols - 1)
-            neighbors.emplace_back(p.col + 1, p.row - 1, img.at<double>(p.row - 1, p.col + 1));
+            neighbors.emplace_back(p.col + 1, p.row - 1, img.at<float>(p.row - 1, p.col + 1));
         if (p.row < img.rows - 1 && p.col < img.cols - 1)
-            neighbors.emplace_back(p.col + 1, p.row + 1, img.at<double>(p.row + 1, p.col + 1));
+            neighbors.emplace_back(p.col + 1, p.row + 1, img.at<float>(p.row + 1, p.col + 1));
         if (p.row > 0 && p.col > 0)
-            neighbors.emplace_back(p.col - 1, p.row - 1, img.at<double>(p.row - 1, p.col - 1));
+            neighbors.emplace_back(p.col - 1, p.row - 1, img.at<float>(p.row - 1, p.col - 1));
         if (p.row < img.rows - 1 && p.col > 0)
-            neighbors.emplace_back(p.col - 1, p.row + 1, img.at<double>(p.row + 1, p.col - 1));
+            neighbors.emplace_back(p.col - 1, p.row + 1, img.at<float>(p.row + 1, p.col - 1));
     }
 
     return neighbors;
